+++ src/widgets/NewRelicErrorBreakdown/components/NewRelicErrorBreakdown.scss
@@ -1,0 +1,6 @@
+.NewRelicErrorBreakdown {
+  overflow: visible;
+  .recharts-tooltip-wrapper {
+    z-index: 10;
+  }
+}
